01-Aug-2018 win folder watcher finished

+++ watchDirRec/src/watchDirRec/watchDirRec.java
@@ -9,9 +9,11 @@ import java.text.SimpleDateFormat;
 import java.io.*;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
+import static java.nio.file.StandardCopyOption.*;
 //import java.time.Duration;
 //import java.time.Instant;
 
+import org.apache.commons.io.FileUtils;
 //apache include
 import org.apache.commons.io.FilenameUtils;
 
@@ -160,15 +162,49 @@ public class watchDirRec
 	                File createdFile = child.toFile();
 	                //System.out.println("extension:"+ext1);
 	                
+	                Collection<String> extn = new ArrayList<String>();
+    				extn.add("rtf");
+    				extn.add("doc");
+    				extn.add("docx");
+	                
 	                // print out event
-	                if(kind.equals(ENTRY_CREATE) && (createdFile.getParentFile().getName().equals("_process")))
-	                {
-	                	DateFormat dateFormat2 = new SimpleDateFormat("dd-MMM-yy HH:mm:ss");
-    	    			String dateString2 = dateFormat2.format(new Date()).toString();
-	                	System.out.println("File("+createdFile.getName()+") created at : "+dateString2);
-	                	docProcessing(child.toString());
-		            }
-	                // if directory is created, and watching recursively, then
+    				if((createdFile.isFile()) && (createdFile.getName().indexOf("~") == -1))
+    				{
+		                if(kind.equals(ENTRY_CREATE) && (createdFile.getParentFile().getName().equals("_process")))
+		                {
+		                	if(FilenameUtils.isExtension(createdFile.getName(),extn))
+							{
+		                		while(!isCompletelyWritten(child.toFile()))
+	    						{}
+								DateFormat dateFormat2 = new SimpleDateFormat("dd-MMM-yy HH:mm:ss");
+								String dateString2 = dateFormat2.format(new Date()).toString();
+								System.out.println("File("+createdFile.getName()+") created at : "+dateString2);
+								docProcessing(child.toString());
+							}
+							else
+							{
+								while(!isCompletelyWritten(child.toFile()))
+	    						{}
+								System.out.println("File("+createdFile.getName()+") is Invalid and moved to \"_log\" folder.");
+								String errPath = createdFile.toString().replace("_process","_log"); 
+								Files.move(createdFile.toPath(),Paths.get(errPath),REPLACE_EXISTING);
+							}
+			            }
+    				}
+    				else if((createdFile.isDirectory()) && (createdFile.getName().indexOf("~") == -1))
+    				{
+    					if(kind.equals(ENTRY_CREATE) && (createdFile.getParentFile().getName().equals("_process")))
+		                {
+    						while(!isCompletelyWritten(createdFile))
+    						{}
+    						Files.delete(createdFile.toPath());
+		                }
+    				}
+//    				else
+//    				{
+//    					Files.delete(createdFile);
+//    				}
+    				// if directory is created, and watching recursively, then
 	                // register it and its sub-directories
 	                if (recursive && (kind == ENTRY_CREATE))
 	                {
@@ -214,16 +250,18 @@ public class watchDirRec
     	 {
     		if((path.indexOf("~") == -1) && (new File(path).exists()))
     		{
-    			String outputPathRtf,inputPathXml,outputPathXml,errPath;
-    			outputPathRtf = path.replace("PDF2XML_Process", "PDF2XML");
-    			outputPathRtf = outputPathRtf.replace("_process", "OUT");
-					
+    			String outputPath,inputPathXml,outputPathXml,errPath,extn;
+    			outputPath = path.replace("PDF2XML_Process", "PDF2XML");
+    			outputPath = outputPath.replace("_process", "OUT");
+				
+    			extn = FilenameUtils.getExtension(path);
+    			
 				//inputPathXml = path;
 				//inputPathXml = path.replace("Process", "Output");
-				inputPathXml = path.replace("rtf", "xml");
+				inputPathXml = path.replace(extn, "xml");
 				//pathXml = pathXml.replace("rtf", "xml");
 				
-				outputPathXml = outputPathRtf.replace("rtf", "xml");
+				outputPathXml = outputPath.replace(extn, "xml");
 				
 				//Path outputPathXml = Paths.get();
 				errPath = path.replace("_process", "_log");
@@ -231,12 +269,12 @@ public class watchDirRec
 				System.out.format("Processing file: "+path+"\n");
 	        	//consoleLog.log("Processing file: "+path+"\n");
 				//System.out.println("input rtf Path:"+path);
-	//			System.out.println("output rtf Path:"+outputPathRtf);
-	//			System.out.println("Error Path:"+errPath);
-	//			System.out.println("input xml path:"+inputPathXml);
-	//			System.out.println("output xml path:"+outputPathXml);
+				//System.out.println("output rtf Path:"+outputPathRtf);
+				//System.out.println("Error Path:"+errPath);
+				//System.out.println("input xml path:"+inputPathXml);
+				//System.out.println("output xml path:"+outputPathXml);
 				
-				File theDir = new File(outputPathRtf).getParentFile();
+				/*File theDir = new File(outputPathRtf).getParentFile();
 				System.out.println("outputPathRtf:"+theDir.exists());
 		    	while (!theDir.exists()) 
 				{
@@ -245,7 +283,7 @@ public class watchDirRec
 				}
 				
 		    	theDir = new File(errPath).getParentFile();
-				System.out.println("outputPathRtf:"+theDir.exists());
+				System.out.println("errPath:"+theDir.exists());
 		    	while (!theDir.exists()) 
 				{
 		    		theDir.mkdir();
@@ -253,7 +291,7 @@ public class watchDirRec
 				}
 		    	
 		    	theDir = new File(inputPathXml).getParentFile();
-				System.out.println("outputPathRtf:"+theDir.exists());
+				System.out.println("inputPathXml:"+theDir.exists());
 		    	while (!theDir.exists()) 
 				{
 		    		theDir.mkdir();
@@ -261,15 +299,15 @@ public class watchDirRec
 				}
 		    	
 		    	theDir = new File(outputPathXml).getParentFile();
-				System.out.println("outputPathRtf:"+theDir.exists());
+				System.out.println("outputPathXml:"+theDir.exists());
 		    	while (!theDir.exists()) 
 				{
 		    		theDir.mkdir();
 		    		TimeUnit.SECONDS.sleep(1);
-				}
+				}*/
 		    	
 				consoleLog.log("input rtf Path:"+path);
-				consoleLog.log("output rtf Path:"+outputPathRtf);
+				consoleLog.log("output rtf Path:"+outputPath);
 				consoleLog.log("Error Path:"+errPath);
 				consoleLog.log("input xml path:"+inputPathXml);
 				consoleLog.log("output xml path:"+outputPathXml);
@@ -318,36 +356,45 @@ public class watchDirRec
 						DateFormat dateFormat2 = new SimpleDateFormat("dd-MMM-yy HH:mm:ss");
     	    			String dateString2 = dateFormat2.format(new Date()).toString();
 						
-    	    			String othFolder1 = System.getProperty ("user.home")+"/Desktop/PDF2XML/01_test/out/"+ new File(path).getName(); 
-    	    			String othFolder2 = System.getProperty ("user.home")+"/Desktop/PDF2XML/01_test/out/"+ new File(outputPathXml).getName();
+    	    			//String othFolder1 = System.getProperty ("user.home")+"/Desktop/PDF2XML/01_test/out/"+ new File(path).getName(); 
+    	    			//String othFolder2 = System.getProperty ("user.home")+"/Desktop/PDF2XML/01_test/out/"+ new File(outputPathXml).getName();
     	    			
     	    			TimeUnit.SECONDS.sleep(3);
     	    			
-						//System.out.println("p.exitValue():"+p.exitValue());
+						System.out.println("p.exitValue():"+p.exitValue());
+						System.out.println("XML status:"+new File(inputPathXml).exists());
 						if((p.exitValue() == -1) || !(new File(inputPathXml).exists()))
 						{
-							System.out.println("File : "+FilenameUtils.getName(path)+",\nMacro Status : FAIL,\n"+"Finished time : "+dateString2+"\n\n");
-							consoleLog.log("File : "+FilenameUtils.getName(path)+",\nMacro Status : FAIL,\n"+"Finished time : "+dateString2+"\n\n");
-							Files.move(Paths.get(path),Paths.get(errPath),StandardCopyOption.REPLACE_EXISTING);
+							if(new File(path).exists())
+							{
+								System.out.println("File : "+FilenameUtils.getName(path)+",\nMacro Status : FAIL,\n"+"Finished time : "+dateString2+"\n\n");
+								consoleLog.log("File : "+FilenameUtils.getName(path)+",\nMacro Status : FAIL,\n"+"Finished time : "+dateString2+"\n\n");
+								Files.move(Paths.get(path),Paths.get(errPath),StandardCopyOption.REPLACE_EXISTING);
+							}
 							//file.delete();
 							//return (false);
 						}
 						else
 						{
-							System.out.println("File : "+FilenameUtils.getName(path)+",\nMacro Status : SUCCESS,\n"+"Finished time : "+dateString2+"\n\n");
-							consoleLog.log("File : "+FilenameUtils.getName(path)+",\nMacro Status : SUCCESS,\n"+"Finished time : "+dateString2+"\n\n");
-							//Files.move(Paths.get(path),Paths.get(outputPathRtf),StandardCopyOption.REPLACE_EXISTING);
-							//Files.move(Paths.get(inputPathXml),Paths.get(outputPathXml),StandardCopyOption.REPLACE_EXISTING);
-							Files.move(Paths.get(path),Paths.get(othFolder1),StandardCopyOption.REPLACE_EXISTING);
-							Files.move(Paths.get(inputPathXml),Paths.get(othFolder2),StandardCopyOption.REPLACE_EXISTING);
+							if(new File(inputPathXml).exists() && new File(path).exists())
+							{
+								System.out.println("File : "+FilenameUtils.getName(path)+",\nMacro Status : SUCCESS,\n"+"Finished time : "+dateString2+"\n\n");
+								consoleLog.log("File : "+FilenameUtils.getName(path)+",\nMacro Status : SUCCESS,\n"+"Finished time : "+dateString2+"\n\n");
+								Files.move(Paths.get(path),Paths.get(outputPath),StandardCopyOption.REPLACE_EXISTING);
+								Files.move(Paths.get(inputPathXml),Paths.get(outputPathXml),StandardCopyOption.REPLACE_EXISTING);
+							}
+							//Files.copy(Paths.get(path),Paths.get(outputPathRtf),StandardCopyOption.REPLACE_EXISTING);
+							//Files.copy(Paths.get(inputPathXml),Paths.get(outputPathXml),StandardCopyOption.REPLACE_EXISTING);
+							//Files.delete(Paths.get(path));
+							//Files.delete(Paths.get(inputPathXml));
 							//file.delete();
 							//return (true);
 						}
 						file.delete();
 						Runtime.getRuntime().exec("taskkill /F /IM WINWORD.EXE");
 						Runtime.getRuntime().exec("taskkill /F /IM wscript.exe");
-						Runtime.getRuntime().exec("attrib -s -h -r "+othFolder1);
-						Runtime.getRuntime().exec("attrib -s -h -r "+othFolder2);
+						//Runtime.getRuntime().exec("cacls "+othFolder1+" -e -p  ");
+						//Runtime.getRuntime().exec("attrib -a -s -h -r "+othFolder2);
 	                }
 	    			else
 	    			{
@@ -373,6 +420,39 @@ public class watchDirRec
     	 return;
     }
 
+    private boolean isCompletelyWritten(File file) throws InterruptedException
+    {
+        Long fileSizeBefore = FileUtils.sizeOf(file);//size(file.toPath());
+        Thread.sleep(3000);
+        Long fileSizeAfter = FileUtils.sizeOf(file);//size(file.toPath());
+
+        //System.out.println("comparing file size " + fileSizeBefore + " with " + fileSizeAfter);
+        if (fileSizeBefore.equals(fileSizeAfter)) 
+        {
+            return true;
+        }
+        System.out.println("Copying.........");
+        return false;
+    }
+    
+    public void setPermission(File file) throws IOException
+    {
+        Set<PosixFilePermission> perms = new HashSet<>();
+        perms.add(PosixFilePermission.OWNER_READ);
+        perms.add(PosixFilePermission.OWNER_WRITE);
+        perms.add(PosixFilePermission.OWNER_EXECUTE);
+
+        perms.add(PosixFilePermission.OTHERS_READ);
+        perms.add(PosixFilePermission.OTHERS_WRITE);
+        perms.add(PosixFilePermission.OTHERS_EXECUTE);
+
+        perms.add(PosixFilePermission.GROUP_READ);
+        perms.add(PosixFilePermission.GROUP_WRITE);
+        perms.add(PosixFilePermission.GROUP_EXECUTE);
+
+        Files.setPosixFilePermissions(file.toPath(), perms);
+    }
+    
     private void initProcessing(String pathString) throws FileNotFoundException, IOException, InterruptedException
     {
     	try
@@ -383,9 +463,39 @@ public class watchDirRec
 			{
 				for (int i = 0; i < listOfFiles.length; i++) 
 				{
+					Collection<String> extn = new ArrayList<String>();
+    				extn.add("rtf");
+    				extn.add("doc");
+    				extn.add("docx");
 					if (listOfFiles[i].isFile() && (listOfFiles[i].getParent().indexOf("_process") != -1)) 
 					{
-						docProcessing(listOfFiles[i].toString());
+						if(FilenameUtils.isExtension(listOfFiles[i].getName(),extn))
+						{
+							while(!isCompletelyWritten(listOfFiles[i]))
+    						{}
+							docProcessing(listOfFiles[i].toString());
+						}
+						else
+						{
+							while(!isCompletelyWritten(listOfFiles[i]))
+    						{}
+							//String errPath = listOfFiles[i].toString().replace("_process","_log");
+							//testPath = errPath;
+//							while(!new File(testPath).getParentFile().exists())
+//							{
+//								File theDir = new File(testPath).getParentFile();
+//								System.out.println("errPath:"+theDir.exists());
+//								System.out.println("nextPath:"+testPath);
+//						    	while (!theDir.exists()) 
+//								{
+//						    		theDir.mkdir();
+//						    		TimeUnit.SECONDS.sleep(1);
+//								}
+//						    	testPath = testPath.substring(0,testPath.lastIndexOf("\\"));
+//							}
+							//Files.move(listOfFiles[i].toPath(),Paths.get(errPath),REPLACE_EXISTING);
+							Files.delete(listOfFiles[i].toPath());
+						}
 					}
 					else if (listOfFiles[i].isDirectory()) 
 					{
@@ -444,6 +554,43 @@ public class watchDirRec
 	        }
 	    }
 	    return false;
+	}
+    
+    public static void delete(File file)throws IOException
+	{ 
+	    if(file.isDirectory())
+	    {
+	    	//directory is empty, then delete it
+	    	if(file.list().length==0)
+	    	{
+	    		file.delete();
+	    		//System.out.println("Directory is deleted : "+ file.getAbsolutePath());
+	    	}
+	    	else
+	    	{
+	    		//list all the directory contents
+        	   String files[] = file.list();
+        	   for (String temp : files) 
+        	   {
+        	      //construct the file structure
+        	      File fileDelete = new File(file, temp);
+        		  //recursive delete
+        	      delete(fileDelete);
+        	   }
+        	   //check the directory again, if empty then delete it
+        	   if(file.list().length==0)
+        	   {
+        		   file.delete();
+        		   //System.out.println("Directory is deleted : "+ file.getAbsolutePath());
+        	   }
+	    	}
+	    }
+	    else
+	    {
+    		//if file, then delete it
+    		file.delete();
+    		//System.out.println("File is deleted : " + file.getAbsolutePath());
+    	}
 	}
     
 	public static void main(String[] args) throws IOException
